fileio: use size_t for pathname lengths

+++ include/blissc/support/fileio.h
@@ -34,7 +34,7 @@ struct fio_pathparts_s {
 };
 typedef struct fio_pathparts_s fio_pathparts_t;
 
-char *file_canonicalname(fioctx_t fio, const char *orig, int origlen, unsigned int *lenp);
+char *file_canonicalname(fioctx_t fio, const char *orig, int origlen, size_t *lenp);
 int file_splitname(fioctx_t fio, const char *orig, int origlen, int canoncialize,
                    fio_pathparts_t *parts);
 int file_combinename(fioctx_t fio, fio_pathparts_t *parts);
